Fixed compiler warning on windows.
git-svn-id: https://svn.opendtect.org/od/trunk@32187 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/Algo/tests/fourier.cc
@@ -93,8 +93,8 @@ bool testForwardCC( bool quiet, const TypeSet<float_complex>& input )
             freqsum += contrib * val;
         }
 
-        reference[idx] = float_complex( freqsum.real(),
-                                        freqsum.imag() );
+        reference[idx] = float_complex( (float) freqsum.real(),
+                                        (float) freqsum.imag() );
     }
 
     Fourier::FFTCC1D transform;
